// src/app/api/model/store-day-lock-request.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StoreDayLockRequest { 
    storeId?: string;
    businessDate?: string;
    note?: string | null;
}